static/resources/.. 파일 추가(정적 요소들 넣어주세요, 이미지, html, css, js, ...) postForm.js 파일 추가 WebMvc설정클래스 삭제
