--- src/main/java/com/project/config/MvcConfig.java
@@ -1,12 +1,0 @@
-package com.project.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class MvcConfig implements WebMvcConfigurer{
-	
-	
-}
