+++ src/main/java/com/project/controller/ModelController.java
@@ -1,0 +1,21 @@
+package com.project.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class ModelController {
+	
+	@GetMapping("/changePasswd")	 
+	public String changePasswdHandler() {
+		return "changePasswd";
+	}
+	
+	@GetMapping("/postForm")	 
+	public String postFormHandler() {
+		return "postForm";
+	}
+	
+	
+}
