+++ src/main/resources/application.properties
@@ -2,4 +2,5 @@ spring.application.name=Career-Board-Project
 
 spring.mvc.view.prefix=/WEB-INF/views/
 spring.mvc.view.suffix=.jsp
+spring.mvc.static-path-pattern=/resources/**
 server.servlet.jsp.init-parameters.development=true
