+++ src/main/resources/static/js/changePasswd.js
